checkers/Sanromolo/requirements.txt
pwntools==4.12.0
requests==2.32.3
